requirements.txt
# URL Shortener - Python Dependencies

# FastAPI Framework
fastapi==0.128.0

# ASGI Server
uvicorn[standard]==0.40.0

# Database ORM
SQLAlchemy==2.0.25

# Environment Variables
python-dotenv==1.0.0

# Form handling
python-multipart==0.0.6

# Templating
jinja2==3.1.3

# Pydantic for validation (already included with fastapi)
pydantic==2.12.5
pydantic_core==2.41.5

# Core dependencies
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
h11==0.16.0
idna==3.11
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0


